task_manager_api/requirements.txt
fastapi>=0.110.0
email-validator>=2.1.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9
slowapi>=0.1.9
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
alembic>=1.13.0
